backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
langchain==0.0.350
langchain-community==0.0.16
langchain-core==0.1.3
langgraph==0.0.20
sentence-transformers==2.2.2
pinecone-client==2.2.4
weaviate-client==3.25.3
firebase-admin==6.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
# OpenAI disabled - using Google AI instead
# openai==1.3.8
google-generativeai==0.3.2
google-cloud-storage==2.10.0
google-cloud-vision==3.4.5
# AWS disabled - using Google Cloud Storage
# boto3==1.34.0
pytesseract==0.3.10
Pillow==10.1.0
python-docx==1.1.0
PyPDF2==3.0.1
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
spacy==3.7.2
bertopic==0.15.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Additional dependencies for Phase 2
chromadb==0.4.18
# Anthropic disabled - using Google AI instead
# anthropic==0.8.1
transformers==4.36.2
umap-learn==0.5.4
hdbscan==0.8.33
python-magic==0.4.27
asyncpg==0.29.0
aioredis==2.0.1
structlog==23.2.0
prometheus-client==0.19.0
matplotlib==3.8.2
seaborn==0.13.0
python-decouple==3.8
jinja2==3.1.2
